+++ lib/lightning/pipeline.ex
@@ -82,10 +82,8 @@ defmodule Lightning.Pipeline do
     end
   end
 
-  def logs_for_run(nil), do: nil
-
   def logs_for_run(%Run{} = run),
-    do: Repo.preload(run, :log_lines) |> Map.get(:log_lines)
+    do: Repo.preload(run, :log_lines) |> Map.get(:log_lines, [])
 
   def assemble_logs_for_run(nil), do: nil